+++ pyxelate/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.4.0"
+__version__ = "2.5.0"
 
 __short_description__ = ("Downsample images to 8-bit pixel art.",)
 __license__ = "MIT"
@@ -8,3 +8,4 @@ __github_username__ = "sedthh"
 from .pal import Pal
 from .pyx import Pyx
 from .vid import Vid
+from .backend import get_backend, is_gpu_available, CUPY_AVAILABLE